+++ MTA10/mods/deathmatch/logic/CSyncDebug.cpp
@@ -28,8 +28,8 @@ CSyncDebug::CSyncDebug ( CClientManager* pManager )
     m_ulLastUpdateTime = CClientTime::GetTime ();
     m_uiLastPacketsSent = m_uiPacketsSent = stats.packetsSent;
     m_uiLastPacketsReceived = m_uiPacketsReceived = stats.packetsReceived;
-    m_uiLastBitsReceived = m_uiBitsReceived = stats.bytesReceived * 8;//  stats.runningTotal [ NS_ACTUAL_BYTES_RECEIVED ] * 8;
-    m_uiLastBitsSent = m_uiBitsSent = stats.bytesSent * 8;//stats.runningTotal [ NS_ACTUAL_BYTES_SENT ] * 8;
+    m_uiLastBitsReceived = m_uiBitsReceived = (uint)stats.bytesReceived * 8;//  stats.runningTotal [ NS_ACTUAL_BYTES_RECEIVED ] * 8;
+    m_uiLastBitsSent = m_uiBitsSent = (uint)stats.bytesSent * 8;//stats.runningTotal [ NS_ACTUAL_BYTES_SENT ] * 8;
 
     m_usFakeLagVariance = 0;
     m_usFakeLagPing = 0;
@@ -181,8 +181,8 @@ void CSyncDebug::OnUpdate ( void )
 
     m_uiPacketsSent = stats.packetsSent;
     m_uiPacketsReceived = stats.packetsReceived;
-    m_uiBitsReceived = stats.bytesReceived * 8;
-    m_uiBitsSent = stats.bytesSent * 8;
+    m_uiBitsReceived = (uint)stats.bytesReceived * 8;
+    m_uiBitsSent = (uint)stats.bytesSent * 8;
 }
